pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loyverse-manager"
version = "0.1.0"
description = "Loyverse automation tools"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{name="Ray Caddick", email="admin@farmyardpark.co.za"}]

# Runtime dependencies - matching prod.txt
dependencies = [
    "python-dotenv>=1.2.0",
    "requests>=2.32.0",
    "selenium>=4.38.0",
    "PyYAML>=6.0.0",
    "pycryptodome>=3.23.0",
    "PyMySQL>=1.1.2",
    "Flask>=3.1.2",
    "reportlab>=4.4.5",
    "phonenumbers>=9.0.19",
    "PyMuPDF>=1.24.0",
    "Pillow>=12.0.0",
    "pyjwt>=2.10.1"
]

# Optional development dependencies
[project.optional-dependencies]
dev = [
    "mypy>=1.18.0",
    "types-requests>=2.32.0",
    "types-reportlab>=4.4.4"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*", "scripts*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.png", "*.pem", "*.db"]

[project.scripts]
add-inventory = "scripts.add_inventory:main"
clear-inventory = "scripts.clear_inventory:main"

[[tool.mypy.overrides]]
module = "fitz"
ignore_missing_imports = true